refactor: Переместить в пакет impl
