+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/UserDbStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.dao.impl.db;
+package ru.yandex.practicum.filmorate.dao.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,7 +20,6 @@ import java.sql.SQLException;
 import java.util.*;
 
 @Repository
-@Qualifier("UserDbStorage")
 @Slf4j
 @RequiredArgsConstructor
 public class UserDbStorage implements UserStorage {
