+++ src/test/java/ru/yandex/practicum/filmorate/storage/UserDbStorageTest.java
@@ -11,8 +11,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import ru.yandex.practicum.filmorate.dao.FriendshipStorage;
 import ru.yandex.practicum.filmorate.dao.UserStorage;
-import ru.yandex.practicum.filmorate.dao.impl.db.FriendshipDbStorage;
-import ru.yandex.practicum.filmorate.dao.impl.db.UserDbStorage;
+import ru.yandex.practicum.filmorate.dao.impl.FriendshipDbStorage;
+import ru.yandex.practicum.filmorate.dao.impl.UserDbStorage;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Friendship;
 import ru.yandex.practicum.filmorate.model.User;
